+++ packages/block-library/src/navigation/test/use-navigation-menu.js
@@ -9,6 +9,12 @@ import { store as coreStore } from '@wordpress/core-data';
  */
 import useNavigationMenu from '../use-navigation-menu';
 
+const BASE_ENTITY = {
+	kind: 'postType',
+	name: 'wp_navigation',
+	id: undefined,
+};
+
 function createRegistryWithStores() {
 	// Create a registry and register used stores.
 	const registry = createRegistry();
@@ -63,37 +69,70 @@ function resolveRecords( registry, menus ) {
 
 function resolveReadPermission( registry, allowed ) {
 	const dispatch = registry.dispatch( coreStore );
-	dispatch.receiveUserPermission( 'create/navigation', allowed );
-	dispatch.startResolution( 'canUser', [ 'read', 'navigation' ] );
-	dispatch.finishResolution( 'canUser', [ 'read', 'navigation' ] );
+	dispatch.receiveUserPermission( 'read/postType/wp_navigation', allowed );
+	dispatch.startResolution( 'canUser', [ 'read', BASE_ENTITY ] );
+	dispatch.finishResolution( 'canUser', [ 'read', BASE_ENTITY ] );
 }
 
 function resolveReadRecordPermission( registry, ref, allowed ) {
 	const dispatch = registry.dispatch( coreStore );
-	dispatch.receiveUserPermission( 'create/navigation', allowed );
-	dispatch.startResolution( 'canUser', [ 'read', 'navigation', ref ] );
-	dispatch.finishResolution( 'canUser', [ 'read', 'navigation', ref ] );
+	dispatch.receiveUserPermission(
+		`read/postType/wp_navigation/${ ref }`,
+		allowed
+	);
+	dispatch.startResolution( 'canUser', [
+		'read',
+		{ ...BASE_ENTITY, id: ref },
+	] );
+	dispatch.finishResolution( 'canUser', [
+		'read',
+		{ ...BASE_ENTITY, id: ref },
+	] );
 }
 
 function resolveCreatePermission( registry, allowed ) {
 	const dispatch = registry.dispatch( coreStore );
-	dispatch.receiveUserPermission( 'create/navigation', allowed );
-	dispatch.startResolution( 'canUser', [ 'create', 'navigation' ] );
-	dispatch.finishResolution( 'canUser', [ 'create', 'navigation' ] );
+	dispatch.receiveUserPermission( 'create/postType/wp_navigation', allowed );
+	dispatch.startResolution( 'canUser', [
+		'create',
+		{ kind: 'postType', name: 'wp_navigation' },
+	] );
+	dispatch.finishResolution( 'canUser', [
+		'create',
+		{ kind: 'postType', name: 'wp_navigation' },
+	] );
 }
 
 function resolveUpdatePermission( registry, ref, allowed ) {
 	const dispatch = registry.dispatch( coreStore );
-	dispatch.receiveUserPermission( `update/navigation/${ ref }`, allowed );
-	dispatch.startResolution( 'canUser', [ 'update', 'navigation', ref ] );
-	dispatch.finishResolution( 'canUser', [ 'update', 'navigation', ref ] );
+	dispatch.receiveUserPermission(
+		`update/postType/wp_navigation/${ ref }`,
+		allowed
+	);
+	dispatch.startResolution( 'canUser', [
+		'update',
+		{ ...BASE_ENTITY, id: ref },
+	] );
+	dispatch.finishResolution( 'canUser', [
+		'update',
+		{ ...BASE_ENTITY, id: ref },
+	] );
 }
 
 function resolveDeletePermission( registry, ref, allowed ) {
 	const dispatch = registry.dispatch( coreStore );
-	dispatch.receiveUserPermission( `delete/navigation/${ ref }`, allowed );
-	dispatch.startResolution( 'canUser', [ 'delete', 'navigation', ref ] );
-	dispatch.finishResolution( 'canUser', [ 'delete', 'navigation', ref ] );
+	dispatch.receiveUserPermission(
+		`delete/postType/wp_navigation/${ ref }`,
+		allowed
+	);
+	dispatch.startResolution( 'canUser', [
+		'delete',
+		{ ...BASE_ENTITY, id: ref },
+	] );
+	dispatch.finishResolution( 'canUser', [
+		'delete',
+		{ ...BASE_ENTITY, id: ref },
+	] );
 }
 
 describe( 'useNavigationMenus', () => {