+++ packages/blocks/src/store/private-actions.js
@@ -51,11 +51,12 @@ export function addBlockBindingsSource( source ) {
 		type: 'ADD_BLOCK_BINDINGS_SOURCE',
 		name: source.name,
 		label: source.label,
-		getValue: source.getValue,
-		setValue: source.setValue,
+		usesContext: source.usesContext,
+		getValues: source.getValues,
 		setValues: source.setValues,
 		getPlaceholder: source.getPlaceholder,
 		canUserEditValue: source.canUserEditValue,
+		getFieldsList: source.getFieldsList,
 	};
 }
 
@@ -68,5 +69,19 @@ export function removeBlockBindingsSource( name ) {
 	return {
 		type: 'REMOVE_BLOCK_BINDINGS_SOURCE',
 		name,
+	};
+}
+
+/**
+ * Add bootstrapped block bindings sources, usually initialized from the server.
+ *
+ * @param {string} source Name of the source to bootstrap.
+ */
+export function addBootstrappedBlockBindingsSource( source ) {
+	return {
+		type: 'ADD_BOOTSTRAPPED_BLOCK_BINDINGS_SOURCE',
+		name: source.name,
+		label: source.label,
+		usesContext: source.usesContext,
 	};
 }