+++ packages/edit-site/src/store/selectors.js
@@ -2,16 +2,19 @@
  * WordPress dependencies
  */
 import { store as coreDataStore } from '@wordpress/core-data';
-import { createRegistrySelector } from '@wordpress/data';
+import { createRegistrySelector, createSelector } from '@wordpress/data';
 import deprecated from '@wordpress/deprecated';
 import { Platform } from '@wordpress/element';
 import { store as preferencesStore } from '@wordpress/preferences';
 import { store as editorStore } from '@wordpress/editor';
+import { store as blockEditorStore } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
 import { unlock } from '../lock-unlock';
+import { TEMPLATE_PART_POST_TYPE } from '../utils/constants';
+import getFilteredTemplatePartBlocks from '../utils/get-filtered-template-parts';
 
 /**
  * @typedef {'template'|'template_type'} TemplateType Template type.
@@ -240,18 +243,46 @@ export function isSaveViewOpened( state ) {
 	return state.saveViewPanel;
 }
 
+function getBlocksAndTemplateParts( select ) {
+	const templateParts = select( coreDataStore ).getEntityRecords(
+		'postType',
+		TEMPLATE_PART_POST_TYPE,
+		{ per_page: -1 }
+	);
+
+	const { getBlocksByName, getBlocksByClientId } = select( blockEditorStore );
+
+	const clientIds = getBlocksByName( 'core/template-part' );
+	const blocks = getBlocksByClientId( clientIds );
+	return [ blocks, templateParts ];
+}
+
 /**
  * Returns the template parts and their blocks for the current edited template.
  *
+ * @deprecated
  * @param {Object} state Global application state.
  * @return {Array} Template parts and their blocks in an array.
  */
 export const getCurrentTemplateTemplateParts = createRegistrySelector(
-	( select ) => () => {
-		return unlock(
-			select( editorStore )
-		).getCurrentTemplateTemplateParts();
-	}
+	( select ) =>
+		createSelector(
+			() => {
+				deprecated(
+					`select( 'core/edit-site' ).getCurrentTemplateTemplateParts()`,
+					{
+						since: '6.7',
+						version: '6.9',
+						alternative: `select( 'core/block-editor' ).getBlocksByName( 'core/template-part' )`,
+					}
+				);
+
+				return getFilteredTemplatePartBlocks(
+					...getBlocksAndTemplateParts( select )
+				);
+			},
+			() => getBlocksAndTemplateParts( select )
+		)
 );
 
 /**